@@ -3,8 +3,8 @@
     "type": "phpbb-extension",
     "description": "Adds the ability to disable and re-enable all installed extensions at once to extension management. An individual selection can be made using checkboxes. In addition, order groups and dependencies can be defined or extensions can be ignored. Additional information is also displayed in the extension management.",
     "homepage": "https://github.com/LukeWCS/ext-mgr-plus",
-    "version": "2.1.0-b6",
-    "time": "2024-10-05",
+    "version": "2.1.0-b7",
+    "time": "2024-10-12",
     "license": "GPL-2.0-only",
     "authors": [
         {
@@ -19,7 +19,7 @@
         }
     ],
     "require": {
-        "php": ">=7.1.3,<8.4.0@dev",
+        "php": ">=7.4.0,<8.4.0@dev",
         "composer/installers": "~1.0.0"
     },
     "extra": {
